pkg/collectors: Move helper functions to utils file
