+++ pkg/collectors/collectors.go
@@ -17,12 +17,7 @@ limitations under the License.
 package collectors
 
 import (
-	"time"
-
-	"regexp"
-
 	"github.com/prometheus/client_golang/prometheus"
-	"k8s.io/api/core/v1"
 	"k8s.io/kube-state-metrics/pkg/metrics"
 )
 
@@ -55,63 +50,4 @@ type MetricFamilyDef struct {
 	Help        string
 	LabelKeys   []string
 	ConstLabels prometheus.Labels
-}
-
-func boolFloat64(b bool) float64 {
-	if b {
-		return 1
-	}
-	return 0
-}
-
-// addConditionMetrics generates one metric for each possible node condition
-// status. For this function to work properly, the last label in the metric
-// description must be the condition.
-func addConditionMetrics(desc *metricFamilyDef, cs v1.ConditionStatus, lv ...string) []*metrics.Metric {
-	ms := []*metrics.Metric{}
-	m, err := metrics.NewMetric(desc.Name, desc.LabelKeys, append(lv, "true"), boolFloat64(cs == v1.ConditionTrue))
-	if err != nil {
-		panic(err)
-	}
-	ms = append(ms, m)
-	m, err = metrics.NewMetric(desc.Name, desc.LabelKeys, append(lv, "false"), boolFloat64(cs == v1.ConditionFalse))
-	if err != nil {
-		panic(err)
-	}
-	ms = append(ms, m)
-	m, err = metrics.NewMetric(desc.Name, desc.LabelKeys, append(lv, "unknown"), boolFloat64(cs == v1.ConditionUnknown))
-	if err != nil {
-		panic(err)
-	}
-	ms = append(ms, m)
-
-	return ms
-}
-
-func kubeLabelsToPrometheusLabels(labels map[string]string) ([]string, []string) {
-	labelKeys := make([]string, len(labels))
-	labelValues := make([]string, len(labels))
-	i := 0
-	for k, v := range labels {
-		labelKeys[i] = "label_" + sanitizeLabelName(k)
-		labelValues[i] = v
-		i++
-	}
-	return labelKeys, labelValues
-}
-
-func kubeAnnotationsToPrometheusAnnotations(annotations map[string]string) ([]string, []string) {
-	annotationKeys := make([]string, len(annotations))
-	annotationValues := make([]string, len(annotations))
-	i := 0
-	for k, v := range annotations {
-		annotationKeys[i] = "annotation_" + sanitizeLabelName(k)
-		annotationValues[i] = v
-		i++
-	}
-	return annotationKeys, annotationValues
-}
-
-func sanitizeLabelName(s string) string {
-	return invalidLabelCharRE.ReplaceAllString(s, "_")
 }
